Create parseInt.js
C programmers and Unix users talk about the standard input channel, often abbreviated to stdin. This is commonly associated with the console device keyboard although operating systems such as Unix allow the user to redirect standard input to almost any data source such as a file, a data communication channel or the output of another program without any need to modify the program in any way. This is a very powerful and useful facility.

The readLine() method belongs to the class BufferedReader. A BufferedReader object must be associated with an InputStreamReader which in turn must be associated with an InputStream object.

Before the conversion can be performed, the input character sequence has to be obtained and stored as a string. A console keyboard user will normally terminate the input by pressing the ENTER key, thus marking out a line of text on console screen. The terminated line of input can be obtained, as a string, using the readLine() input method. This will internally detect and remove the terminating return character and construct a properly formed string. The sequence of events is:

Variance refers to how subtyping between more complex types relates to subtyping between their components. “More complex types” here refers to higher-level structures like containers and functions. So, variance is about the assignment compatibility between containers and functions composed of parameters that are connected via a Type Hierarchy. It allows the safe integration of parametric and subtype polymorphism.